Add memory free

+++ libs/allocator/utils.h
@@ -2,6 +2,7 @@
 #define ALLOCATOR_UTILD_H
 
 #include <stddef.h>
+#include <stdint.h>
 
 long int get_system_page_size();
 
@@ -11,6 +12,6 @@ size_t get_correct_pages_number(size_t memory_size);
 
 size_t get_memory_size_for_new_area(size_t memory_size);
 
-uint32_t crc8(uint8_t* data, size_t length);
+uint8_t crc8(uint8_t* data, size_t length);
 
 #endif
